- Make the library header only
