+++ simpleMath/include/Slice.h
